Utilities: Don't inject events if screen OFF
Strangely enough, when running the fbdev emulation on top of a KMS/DRM
driver, when calling SDL_PushEvent() the screen turns ON by itself if it
was OFF before.

Work around that by only sending the repaint event if the screen is ON.
It's not like we have to show anything when the screen is OFF anyway.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>